requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn>=0.23.2

# Database
redis>=4.6.0
chromadb>=0.4.18

# Machine Learning & NLP
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.2
dspy-ai>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.4.2
numpy>=1.24.0
pytest>=7.4.0
